refactor: add PDFViewerContainer and pdfViewer constants
