+++ frontend/src/constants/pdfViewer.ts
@@ -1,0 +1,5 @@
+import { pdfjs } from 'react-pdf';
+
+export const FIRST_PAGE_INDEX = 1;
+export const LAST_PAGE_INDEX = 3;
+export const PDF_WORKER_SRC = `//cdnjs.cloudflare.com/ajax/libs/pdf.js/${pdfjs.version}/pdf.worker.min.js`;
