+++ frontend/src/pages/Help/PDFViewer/index.tsx
@@ -1,1 +1,1 @@
-export { default } from './PDFViewer';
+export { default } from './PDFViewerContainer';
